segemenrted code for easier debugging
